tools/connectivity/requirements.txt
paho-mqtt>=2.1.0
pyserial>=3.5
